refactor: Replace TypeSafeServiceProvider with ServiceProvider in ContainerBuilder and update related documentation, enhancing clarity and type safety in dependency injection.
